Fix installation status check for install scripts

+++ uninstall.sh
@@ -6,7 +6,7 @@ if [ "$(id -u)" -ne 0 ]; then
 fi
 
 MODULES=$(lsmod | grep '^xone_' | cut -d ' ' -f 1 | tr '\n' ' ')
-INSTALLED=$(dkms status xone | tr -s ',:' ' ' | cut -d ' ' -f 2)
+INSTALLED=$(dkms status xone | head -n 1 | tr -s ',:' ' ' | cut -d ' ' -f 2)
 SOURCE="/usr/src/xone-$INSTALLED"
 BLACKLIST='/etc/modprobe.d/xone-blacklist.conf'
 
@@ -18,9 +18,10 @@ fi
 
 if [ -n "$INSTALLED" ]; then
     echo "Uninstalling xone $INSTALLED..."
-    dkms remove --all "xone/$INSTALLED"
-    rm -r "$SOURCE"
-    rm -f "$BLACKLIST"
+    if dkms remove --all "xone/$INSTALLED"; then
+        rm -r "$SOURCE"
+        rm -f "$BLACKLIST"
+    fi
 else
-    echo 'Driver is not installed!'
+    echo 'Driver is not installed!' >&2
 fi
